feat: sidebar drawer preview

+++ libs/elbe/drawer/src/lib/elb-drawer.ts
@@ -107,6 +107,8 @@ export class ElbDrawer {
   }
 
   private async destroy() {
+    if (!this.cupertino?.isPanePresented()) return;
+
     await this.cupertino?.destroy({ animate: this.animate() });
   }
 }
